server/requirements.txt
pyyaml>=5.2
pymysql>=0.9.3